Added tests to confirm inventory management is transactional.
Added two tests. One test confirms that the OrderConfirmation
service successfully decrements the product stock quantity
by the order amount. The second forces a failure within the
transaction to save the order and insures that the product
stock quantity remains the same as it was prior to the order
creation attempt.