Added test to confirm record locks honor full transaction boundaries.
Added a threaded test to try and simulate the loss of a transaction
due to a competing process acquiring a write lock before the
transaction can complete. The test mirrors the concurrency code
used in the OrderConfirmationService to insure that the product
quantity is protected by the ActiveRecord transaction handling.